Add the base for the rbintegrations extension.
The rbintegrations extension will be home to a set of integrations for
third-party services, usable by Review Board 3.0+. We'll be depending on
this in Review Board 3.0, so that out of the box, users will have
immediate access to these integrations. It's a separate package so that
fixes and features can be released independent of Review Board releases.

There aren't any integrations as of this change, just the foundation for
the package and the extension. They will come in future integrations.

Reviewed at https://reviews.reviewboard.org/r/8164/