Add integration for CircleCI.
This change adds a new integration which allows people to build prospective
changes via CircleCI. This requires the (beta) CircleCI 2.0.

This is similar in many ways to the Travis CI integration, but is somewhat
different in its configuration. The major difference is that while Travis
allows us to override the config within our API request, Circle does not. This
means that users will have to add Review Board support within their
`.circleci/config.yml` file.

When a config matches, this integration will trigger a build of the
"reviewboard" job (as defined in the user's config) and pass along various bits
of data in the build parameters.

Testing done:
Posted a variety of changes which should either fail or pass the build, and
saw the results reported correctly.

Reviewed at https://reviews.reviewboard.org/r/8951/