Fix CI integrations failing when URLAvatarService is disabled
If URLAvatarService is disabled, both the TravisCI and CircleCI integrations
will fail to create their service user. This is due to attempting to access the
URLAvatarService without checking if it is enabled.

Here the fix is to check if it's enabled, then proceed with using the service.

Testing Done:
Ran the unit tests for both travisci and circleci integrations, as
well as manual testing with my jenkinsci code.

Reviewed at https://reviews.reviewboard.org/r/9721/