Fix dropdown z-indexes for Asana and Trello fields.
In the case where a review request that had file attachments also had
the Asana or Trello integration fields active, the event overlay on the
file attachment would be on top of the dropdown. This overlay is
invisible, so it was just catching events meant for thee dropdown.
This change adds some CSS rules to put those at the correct z-index.

Testing done:
Added a file attachment and verified that I could now select items in
the Asana and Trello dropdown menus that happened to be at the same
coordinates as the file attachment.

Reviewed at https://reviews.reviewboard.org/r/9801/