Update for changes in Travis CI's webhook payload format.
Travis CI has made some changes to their webhook payload. These are
currently effective on travis-ci.com, and will likely be making it to
travis-ci.org soon.

The biggest of these is that the `global_env` key is completely gone.
There's a top-level `env` key, but that's in a weird hard-to-parse
format. I've changed it to pull the relevant keys out of the matrix data
instead.

Second, I was encountering a new `errored` state for some builds. I've
added handling for this and logging for any future states which we don't
handle.

Testing Done:
Posted some changes that triggered Travis CI builds and saw the webhook
handler work correctly.

Reviewed at https://reviews.reviewboard.org/r/10970/