Fix Travis CI username formatting for github public repos.
The calculation of the repository path was using the wrong key for
public repositories on GitHub (for individual users, not an
organization). This change fixes it.

Testing Done:
Was able to configure and use Travis CI with a non-organization
public repository on GitHub.

Reviewed at https://reviews.reviewboard.org/r/10972/