Include parent diffs in Travis CI job payload.
The Travis CI implementation wasn't serializing parent diffs into the
job payload. This sometimes resulted in builds which would wait and
eventually time out as it tried to apply a patch that didn't work
and waited for user input (why patch thinks it's on a tty is beyond me).

This change adds serialization of the parent diff, if necessary. We
don't have any facilities for doing this easily, so it basically
replicates the logic that exists for normal diffs.

Testing Done:
Triggered successful Travis CI builds for changes that had parent diffs,
and verified that changes without parent diffs were unaffected.

Reviewed at https://reviews.reviewboard.org/r/10973/