Don't use `git checkout --unshallow` for full-repo Travis CI builds.
Some people use `git: depth: False` in their Travis CI config, most
commonly because tests may need full access to the repository history.
In this case, we were still including `git checkout --unshallow` in the
build script, which wasn't fatal but produced ugly warnings. This change
checks for a full-repo checkout in the Travis config and only adds the
checkout if necessary.

Testing Done:
Created a Travis integration that included `git: depth: False` in the
config. Saw that prior to this change, an ugly error was shown in the
build log, and after everything was fine.

Reviewed at https://reviews.reviewboard.org/r/10974/