Move Travis CI checkout and patching to `before_install` section.
In repos where dependencies are changing between branches, having the
checkout and patch steps run as part of the Travis CI `script` would
sometimes fail (for example, attempting to do builds of djblets on
release-1.0.x failed in `install` because it was trying to get
django-evolution 2.0, as required in the master branch).

Doing the checkout and patch in `before_install` makes a lot more sense,
because then we've got the full change in the working directory before
doing anything.

Testing Done:
Ran a build of a change from djblets-1.0.x and saw that it got past the
initial install/setup steps, where before it didn't.

Reviewed at https://reviews.reviewboard.org/r/10976/