Fix Jenkins compatibility with Review Board 3.0.
The new Jenkins integration is mostly fine running on top of old Review
Board versions, but there was one utility class that was renamed. This
change fixes it to use the correct import for both 3.0 and 4.0.

Testing Done:
Was able to load the master branch of the rbintegrations extension on
Review Board 3.0.x.

Reviewed at https://reviews.reviewboard.org/r/10980/