Change 'Ship It!' color 'success' to 'good' for chat integrations
Currently, when reviewers select the "Ship It!" option, the notification sent
to Slack has a bar that is colored grey. I found that a value of 'success' is
passed to color. Upon investigation it seems that Slack and Mattermost doesn't
understand what color corresponds to 'success'. However, it knows 'good'
corresponds to green.

Evidence that Mattermost uses ‘good’ and not ‘success’:
https://mattermost.atlassian.net/browse/MM-3181
https://mattermost.com/blog/getting-started-with-mattermost-integrations/

Evidence that Slack uses ‘good’:
https://api.slack.com/reference/messaging/attachments#fields

Testing Done:
* I ran the unit tests on Slack and Mattermost. Two unrelated tests are failing
  because they cannot find `logo.png` at `/var/folders/...`. See #integrations
* Confirmed that "Ship It!" message has a green bar instead of a grey one in
  Slack and Mattermost for Review Board version 3.0 and 4.0 after the change

Reviewed at https://reviews.reviewboard.org/r/11211/