Add manual trigger run and retry for JenkinsCI integration.
Add manual trigger to JenkinsCI.

This change adds the ability for Integrations to run JenkinsCI manually.
A checkbox is added to the JenkinsCI form that indicates if JenkinsCI should
only run manually.

The password form label has also been updated to emphasize using an API token
instead of the user's password since newer versions of Jenkins only works with
API tokens.

This change also fixes the Jenkins API failing for Review Board 3 because
`request.open()` was not valid.

Based on work by David Trowbridge: [/r/11130](https://reviews.reviewboard.org/r/11130/)

Testing Done:
Manually tested on Review Board 3. Jenkins was set up with Docker and
connected to a GitHub repository and Review Board 3. A Jenkins build was
not run when the manually run checkbox is unchecked and is run when the
field is checked. Clicking on the 'Run' button runs the Jenkins build.

Verified the disabled/enabled run manually field by testing on Review Board
3.0.18, 3.0.19 and 4.0.x.

Verified retry functionality when state is in ERROR and TIMEOUT states.

Unit tests to verify that a build is not run when the manually run
configuration is checked and to verify that a build is run when the signal is
emitted.

Reviewed at https://reviews.reviewboard.org/r/11225/