Add an integration for Discord.
This works like the Slack and Mattermost integrations, allowing
Review Board to notify Discord when review requests are created,
updated, or closed, and when there are new discussions.

The integration makes use of Discord's Slack compatibility endpoint,
which allows us to reuse much of the Slack payload code. To help with
this, the message building portion of the `notify()` function in the
Slack integration was split out into a new `build_slack_message()`
function that can be more easily reused.

Testing Done:
New and existing unit tests pass for Review Board 3.x and 4.0.

Tested manually in Review Board 3.0.x and 4.0.