Fix CircleCI builds not running on personal GitHub repositories.
CircleCI pipelines were not running for personal GitHub repositories
due to the GitHub username was being extracted from an invalid key in
`extra_data`. It's been updated to use the username in the
`HostingServiceAccount` associated with the repository.

Testing Done:
Manual testing: A personal public and private repository was set up and
integrated with CircleCI. They are confirmed to be running without using
Review Board. Then, a review request was made for both the repositories.
Using the CircleCI web UI, the pipelines were triggered.

Unit tests were created to verify that the `username` is being pulled
from the correct place for the 4 different GitHub repository plans.
These were ran on Review Board 3.0 and Review Board 4.0 and their
supported versions of Python.

Reviewed at https://reviews.reviewboard.org/r/11190/