Switch the test runner to be a Python script.
A previous change introduced a `runtests.sh` script, which ran the unit
tests using `rbext`. Doing this as a shell script was simple, but
unfortunately it made it hard to run with different versions of Python.

This change switches this to a Python script, which can easily be run
with any version of Python. It wraps the `RBExt()` class, passing in the
required arguments. So long as `rbext` is installed with that version of
Python, it will work.

Testing Done:
Ran the unit test suite with multiple versions of Python.

Reviewed at https://reviews.reviewboard.org/r/11664/