Fix Python 3 compatibility when sending messages to Slack or Mattermost.
The code in Slack (which Mattermost uses as well) to send messages had a
string type problem. A Unicode string was being passed as the payload
body to `urlopen()`, which wasn't valid on Python 3. While the exception
was caught, it meant that no messages could be sent to either service.

This change fixes the string type, and updates all the relevant unit
tests to perform checks on the string types, headers, and payload
contents.

The unit test updates also simplify the tests, introducing common
utility functions to each suite that performs all the common tests.

Testing Done:
Unit tests pass on Python 2 and 3.

Reviewed at https://reviews.reviewboard.org/r/11665/