Add parameter REVIEWBOARD_REVIEW_BRANCH to jenkinsci
If a user has one jenkins job for all branches it is necessary to get
be able to pass through the branch name in order to check out the
correct branch.

Testing Done:
Created a new review request and started jenkins job for that request.
Saw that the jenkins received the correct branch.

Reviewed at https://reviews.reviewboard.org/r/11747/