Always replace {branch} and replace / with _
{branch} shouldn't be a Git-only variable. Any version system can use
this. If a repository name contains / we need to convert it into another
character because Jenkins do not allow / as a job name. So let's use _
as a replacement.

Testing Done:
Configured a jenkins integration with job name ```{branch}_{repository}```.
The repository name contains multiple ```/```. Started a review request
and saw that it correctly started a job with replaced names.

Reviewed at https://reviews.reviewboard.org/r/11758/