Add manual trigger run and retry for CircleCI integration.
This change adds the ability for Integrations to run CircleCI manually.
A checkbox is added to the CircleCI form that indicates if CircleCI should only
run manually.

Testing Done:
Manual testing by adding a private GitHub repository with CircleCI integration,
checking and unchecking the "manually run" form checkbox to ensure that no
automated runs and done when a review request is made, and clicking the "Run"
button on the review request page.

Verified the disabled/enabled run manually field by testing on Review Board
3.0.18, 3.0.19 and 4.0.x.

Verified retry functionality when state is in ERROR and TIMEOUT states.

Unit tests have been written to verify that when the configuration is set,
the CircleCI request is not made. Another test to verify that a build is
run when the signal is emitted.

Reviewed at https://reviews.reviewboard.org/r/11208/