Cap the asana dependency on Python 2.7.
rbintegrations 2.x had an uncapped dependency on `asana`, which made it
difficult to install on Python 2.7, as Asana 3.x no longer supports
Python 2.7.

This change adds a cap to the dependency on 2.7, and bumps up the
minimum version to 3.x on Python 3.

Testing Done:
Unit tests passed on Python 2.7 and 3.6-3.10.

Verified the version of asana used in each of those versions.

Reviewed at https://reviews.reviewboard.org/r/12858/