Fix UI for adding API credentials for Trello integration.
Trello increased the size of their API tokens, breaking our UI.
According to their documentation for the change, API keys and tokens are
expected to be opaque values, so we shouldn't have been enforcing any
kind of limits on these anyway.

This also updates the help text for new URLs and names in the Trello UI.

Testing Done:
Registered a trello integration and was able to link to cards in the
field.

Reviewed at https://reviews.reviewboard.org/r/13384/