Add an integration for Microsoft Teams.
This adds an integration for Microsoft Teams. Similar to our other chat
service integrations, this sends messages to channels in Microsoft Teams when
review requests are created, updated, and reviewed.

We use Microsoft's [Adaptive Card](https://learn.microsoft.com/en-us/microsoftteams/platform/webhooks-and-connectors/how-to/connectors-using?tabs=cURL#send-adaptive-cards-using-an-incoming-webhook) format for the messages. Unlike our other chat
integrations, with MS Teams there's unfortunately no ability to add an accent
colour to the message (typically we use yellow, green, and red accents as
a visual indicator for the tone of a message). The styling of the messages
also differ between the MS teams desktop and mobile apps, but this is a
choice by Microsoft and out of our control.

Based on work by Hailan Xu (/r/11227/).

Testing Done:
Ran unit tests.

Manually tested on RB5, RB6 and RB7:
- Creating the integration.
- New review requests, with/without diffs and files.
- Updates to review requests.
- Reviews with/without comments, issues, ship its.
- Replies.
- Tested that text which contains Markdown characters
  gets properly escaped when being formatted in a link
  in a message.
- Tested that URLs which contain ( or ) gets properly
  escaped when being formatted in a link in a message.

Reviewed at https://reviews.reviewboard.org/r/13742/