Add control over build timeouts for all CI integrations.
This introduces a new timeout field for configuring the timeout period
for builds on CircleCI, Jenkins, and Travis-CI. This is used to ensure
that if a build takes too long, it can be retried by the review request
owner.

The field uses `AmountSelectorWidget`, configured to let users specify a
number and select "seconds", "minutes", "hours", or "Never" for the
timeout.

Testing Done:
Unit tests passed.

Manually configured the timeouts and verified the configuration in
Status Updates.

Reviewed at https://reviews.reviewboard.org/r/13868/