Add new icons for MS Teams and Discord, and improve asset building.
We've updated the icons for both MS Teams and Discord in our CDN to
better fit into the rounded avatars in the messages. This change updates
the assets timestamp to force use of the new images and to specify the
path for the new MS Teams assets.

In order to properly leverage the assets timestamp and base URL for the
logo and trophies without having to overwrite all the attributes, we now
define `logo_url` and `trophy_urls` as cached properties instead of
attributes. This allows them to be computed dynamically based on any
overridden asset timestamp/base URLs defined by subclasses.

Testing Done:
Tested the new download assets and generated URLs.

This will require some testing in production.

Reviewed at https://reviews.reviewboard.org/r/13914/