Fix links in chat notifications for review replies.
Chat notifications were blindly calling `review.get_absolute_url()` to
get the URL to link to for the notification. Unfortunately, for
review replies, this doesn't work because we don't put an anchor for
those into the page.

This change fixes it so if we have a comment (which we use for the
notification text), we open directly to that comment, and if we don't,
we link to the parent review.

Testing Done:
- Ran unit tests.
- Verified that generated links worked correctly.

Reviewed at https://reviews.reviewboard.org/r/13943/