Fix spurious logging triggered from Trello searches.
While testing out the Trello integration, I noticed that if I typed in a
search and then backspaced it out, we'd get errors in the log about
invalid API requests. The problem is that our selectize-based search
triggers an API request every time it changes, even when the input is
emptied out, but the Trello API requires the search text to be at least
1 character long.

This change makes it so if the search query is empty, we just directly
return an empty list to the client instead of trying to make an API
request to Trello.

Testing Done:
Emptied out my search input and saw that the search results drop-down
disappeared without any ugly errors being logged on the server.

Reviewed at https://reviews.reviewboard.org/r/14185/