Fix imports for legacy condition support.
The recent work on extensible condition support in rbintegrations
introduced a backwards-compatible implementation for earlier versions of
Review Board, but this ended up introducing an import that wasn't
present in that module on older versions.

That import was only needed for type hints anyway, so it's been moved
out into a `TYPE_CHECKING` block and updated to use the correct module
and class for the supported releases.

Testing Done:
Tested on Review Board 7.0.2, and verified I no longer hit import errors.

Reviewed at https://reviews.reviewboard.org/r/14383/