Correct typing in the conditions compatibility file.
We recently added a conditions compatibility file to Review Bot
(/r/14705/), and in those reviews we decided on a new best pattern
for compatibility imports that keep the type checker happy. This change
updates the compatibility file here to use that pattern.

The pattern ensures that we're giving the type checker one path to check
based on the upstream, making sure we're always typing against the
formal API instead of the compatibility stubs.

Testing Done:
- Saw things are typed properly when I import and use the conditions
  field into another file.
- Looked at an integration config page on an RB 7.1 server.

Reviewed at https://reviews.reviewboard.org/r/14732/